# laravel/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Swagger with Laravel",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/products": {
            "get": {
                "tags": [
                    "Products"
                ],
                "summary": "Get list of products",
                "description": "Returns list of products",
                "operationId": "getProductsList",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Product"
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "No products found"
                    }
                }
            },
            "post": {
                "tags": [
                    "Products"
                ],
                "summary": "Store a new product",
                "description": "Stores a new product",
                "operationId": "storeProduct",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Product"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Product created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Product"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error"
                    }
                }
            }
        },
        "/api/products/{id}": {
            "get": {
                "tags": [
                    "Products"
                ],
                "summary": "Get a specific product",
                "description": "Returns a specific product",
                "operationId": "getProductById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the product",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Product"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Product not found"
                    }
                }
            },
            "put": {
                "tags": [
                    "Products"
                ],
                "summary": "Update a specific product",
                "description": "Updates a specific product",
                "operationId": "updateProduct",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the product",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Product"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Product updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Product"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Product not found"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Products"
                ],
                "summary": "Delete a specific product",
                "description": "Deletes a specific product",
                "operationId": "deleteProduct",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the product",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Product deleted successfully"
                    },
                    "404": {
                        "description": "Product not found"
                    }
                }
            }
        },
...
        "/users": {
            "post": {
                "tags": [
                    "User"
                ],
                "summary": "Create a new User",
                "description": "Store a newly created resource in storage.",
                "operationId": "a1fce870c566eb5fc0cf3178ec5d2fed",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/User"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "User created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/User"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/users/{id}": {
            "get": {
                "tags": [
                    "User"
                ],
                "summary": "Get a specific User by ID",
                "description": "Display the specified resource.",
                "operationId": "36a33ff774d5cba33c039dec2c3e0287",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "User ID",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/User"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "User not found"
                    }
                }
            },
            "put": {
                "tags": [
                    "User"
                ],
                "summary": "Update a specific User by ID",
                "description": "Update the specified resource in storage.",
                "operationId": "b9091397c8b25f12c6adb74be6ce3a5a",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "User ID",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/User"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "User updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/User"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "User not found"
                    }
                }
            },
            "delete": {
                "tags": [
                    "User"
                ],
                "summary": "Delete a specific User by ID",
                "description": "Remove the specified resource from storage.",
                "operationId": "fa56cffde745d3f152f95cbacd936c0b",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "User ID",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "User deleted successfully"
                    },
                    "404": {
                        "description": "User not found"
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "Artisan": {
                "properties": {
                    "id": {
                        "description": "Artisan ID",
                        "type": "integer",
                        "format": "int64"
                    },
                    "created_at": {
                        "description": "Artisan creation timestamp",
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "description": "Artisan update timestamp",
                        "type": "string",
                        "format": "date-time"
                    },
                    "user": {
                        "$ref": "#/components/schemas/User"
                    }
                },
                "type": "object"
            },
            "Consumer": {
                "properties": {
                    "id": {
                        "description": "Consumer ID",
                        "type": "integer",
                        "format": "int64"
                    },
                    "created_at": {
                        "description": "Consumer creation timestamp",
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "description": "Consumer update timestamp",
                        "type": "string",
                        "format": "date-time"
                    },
                    "user": {
                        "$ref": "#/components/schemas/User"
                    },
                    "order": {
                        "$ref": "#/components/schemas/Order"
                    }
                },
                "type": "object"
            },
            "DeliveryPersonnel": {
                "properties": {
                    "id": {
                        "description": "DeliveryPersonnel ID",
                        "type": "integer",
                        "format": "int64"
                    },
                    "created_at": {
                        "description": "DeliveryPersonnel creation timestamp",
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "description": "DeliveryPersonnel update timestamp",
                        "type": "string",
                        "format": "date-time"
                    },
                    "user": {
                        "$ref": "#/components/schemas/User"
                    }
                },
                "type": "object"
            },
            "Order": {
                "properties": {
                    "id": {
                        "description": "Order ID",
                        "type": "integer",
                        "format": "int64"
                    },
                    "orderDate": {
                        "description": "Order date",
                        "type": "string",
                        "format": "date-time"
                    },
                    "orderStatus": {
                        "description": "Order status",
                        "type": "string"
                    },
                    "deliveryAddress": {
                        "description": "Delivery address",
                        "type": "string"
                    },
                    "consumer_id": {
                        "description": "ID of the associated consumer",
                        "type": "integer",
                        "format": "int64"
                    },
                    "created_at": {
                        "description": "Order creation timestamp",
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "description": "Order update timestamp",
                        "type": "string",
                        "format": "date-time"
                    },
                    "deliveryPersonnel": {
                        "$ref": "#/components/schemas/DeliveryPersonnel"
                    },
                    "consumer": {
                        "$ref": "#/components/schemas/Consumer"
                    },
                    "orderProducts": {
                        "description": "Related OrderProducts",
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/OrderProduct"
                        }
                    }
                },
                "type": "object"
            },
            "OrderProduct": {
                "properties": {
                    "id": {
                        "description": "OrderProduct ID",
                        "type": "integer",
                        "format": "int64"
                    },
                    "quantity": {
                        "description": "OrderProduct quantity",
                        "type": "integer"
                    },
                    "order": {
                        "$ref": "#/components/schemas/Order"
                    },
                    "product": {
                        "$ref": "#/components/schemas/Product"
                    },
                    "created_at": {
                        "description": "OrderProduct creation timestamp",
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "description": "OrderProduct update timestamp",
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
            "Product": {
                "properties": {
                    "id": {
                        "description": "Product ID",
                        "type": "integer",
                        "format": "int64"
                    },
                    "name": {
                        "description": "Product name",
                        "type": "string"
                    },
                    "description": {
                        "description": "Product description",
                        "type": "string"
                    },
                    "pricePerPiece": {
                        "description": "Product price per piece",
                        "type": "number",
                        "format": "float"
                    },
                    "minOrder": {
                        "description": "Minimum order quantity",
                        "type": "integer"
                    },
                    "type": {
                        "description": "Product type",
                        "type": "string",
                        "enum": [
                            "Sugar",
                            "Salt"
                        ]
                    },
                    "childType": {
                        "description": "Product child type",
                        "type": "string"
                    },
                    "images": {
                        "description": "Product images",
                        "type": "json"
                    },
                    "ratings": {
                        "description": "Related Ratings",
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Rating"
                        }
                    },
                    "artisan": {
                        "$ref": "#/components/schemas/Artisan"
                    },
                    "orderProducts": {
                        "description": "Related OrderProducts",
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/OrderProduct"
                        }
                    },
                    "created_at": {
                        "description": "Product creation timestamp",
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "description": "Product update timestamp",
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
            "Rating": {
                "properties": {
                    "id": {
                        "description": "Rating ID",
                        "type": "integer",
                        "format": "int64"
                    },
                    "rating": {
                        "description": "Numeric rating value",
                        "type": "integer"
                    },
                    "ratingType": {
                        "description": "Type of rating (DeliveryPersonnel, Product, Artisan)",
                        "type": "string",
                        "enum": [
                            "DeliveryPersonnel",
                            "Product",
                            "Artisan"
                        ]
                    },
                    "created_at": {
                        "description": "Rating creation timestamp",
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "description": "Rating update timestamp",
                        "type": "string",
                        "format": "date-time"
                    },
                    "delivery_personnel": {
                        "$ref": "#/components/schemas/DeliveryPersonnel"
                    },
                    "product": {
                        "$ref": "#/components/schemas/Product"
                    },
                    "artisan": {
                        "$ref": "#/components/schemas/Artisan"
                    }
                },
                "type": "object"
            },
            "User": {
                "properties": {
                    "id": {
                        "description": "User ID",
                        "type": "integer",
                        "format": "int64"
                    },
                    "email": {
                        "description": "User email",
                        "type": "string"
                    },
                    "first_name": {
                        "description": "User first name",
                        "type": "string"
                    },
                    "last_name": {
                        "description": "User last name",
                        "type": "string"
                    },
                    "description": {
                        "description": "User description",
                        "type": "string"
                    },
                    "adresse": {
                        "description": "User address",
                        "type": "string"
                    },
                    "phone_number": {
                        "description": "User phone number",
                        "type": "string"
                    },
                    "user_type": {
                        "description": "User type",
                        "type": "string",
                        "enum": [
                            "Artisan",
                            "Consumer",
                            "DeliveryPersonnel"
                        ]
                    },
                    "created_at": {
                        "description": "User creation timestamp",
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "description": "User update timestamp",
                        "type": "string",
                        "format": "date-time"
                    },
                    "deliveryPersonnel": {
                        "$ref": "#/components/schemas/DeliveryPersonnel"
                    },
                    "artisan": {
                        "$ref": "#/components/schemas/Artisan"
                    },
                    "consumer": {
                        "$ref": "#/components/schemas/Consumer"
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}